[Wip] - Moulitest errors corrections. 18 left.

+++ libft/includes/libftprintf.h
@@ -235,7 +235,7 @@ void				print_int(t_pf_item *pfi, intmax_t num);
 void				print_unsigned_int(t_pf_item *pfi, intmax_t num);
 void				print_long(t_pf_item *pfi, long num);
 void				print_address(t_pf_item *pfi, long int num);
-void				print_hex(t_pf_item *pfi, int n);
+void				print_hex(t_pf_item *pfi, intmax_t n);
 void				print_hex_long(t_pf_item *pfi, long long int num);
 void				print_hex_long_prefix(t_pf_item *pfi);
 void				print_oct(t_pf_item *pfi, intmax_t n);
